Move CLI tests into their own file.
Move the recently added test which do not touch `index.js` code
directly into `nyc-bin.js` and their corresponding fixtures to
`test/fixtures/cli`.
