+++ test/fixtures/cli/half-covered-failing.js
