+++ test/fixtures/cli/half-covered.js
