+++ test/fixtures/cli/package.json
@@ -1,0 +1,3 @@
+{
+  "private": true
+}
