+++ test/fixtures/cli/run-npm-test-recursive/package.json
