+++ test/fixtures/cli/run-npm-test/half-covered.js
@@ -1,0 +1,9 @@
+var a = 0
+
+a++
+
+if (a === 0) {
+  a++;
+  a--;
+  a++;
+}
