+++ test/fixtures/cli/run-npm-test/package.json
