--- test/fixtures/run-npm-test-recursive/half-covered.js
@@ -1,1 +1,0 @@
-../half-covered.js
