--- test/fixtures/run-npm-test/half-covered.js
@@ -1,1 +1,0 @@
-../half-covered.js
